сделала обновление спецификации и решила проблему с неправильным подсчетом общей суммы для проудкта и добавил в приложение экран для локаций и начала делать экран для создания локации

+++ features/application-sections/crm-section/crm-feature/crm-impl/src/commonMain/kotlin/model/SpecificResponseModel.kt
@@ -24,7 +24,7 @@ data class SpecItemModel (
         val product_id: Int?,
         val price_id: Int? = null,
         val block: String?,
-        val count: Int?,
+        val count: Float?,
         val price: Double?,
         val nds: Int? = null,
         val text: String? = null,
